backend/requirements.txt
flask
langchain
openai
langchain-openai
python-dotenv
